refactor: 불필요한 추상화 다시 복구
